Add support for systemd socket activation

+++ cmd/gitsign-credential-cache/main.go
@@ -23,6 +23,7 @@ import (
 	"path/filepath"
 	"syscall"
 
+	"github.com/coreos/go-systemd/activation"
 	"github.com/pborman/getopt/v2"
 
 	"github.com/sigstore/gitsign/internal/cache/service"
@@ -32,6 +33,7 @@ import (
 var (
 	// Action flags
 	versionFlag = getopt.BoolLong("version", 'v', "print the version number")
+	systemdFlag = getopt.BoolLong("socket-activation", 's', "use systemd socket activation")
 )
 
 func main() {
@@ -47,31 +49,53 @@ func main() {
 		os.Exit(0)
 	}
 
-	user, err := os.UserCacheDir()
-	if err != nil {
-		log.Fatalf("error getting user cache directory: %v", err)
-	}
+	var connChan = make(chan net.Conn)
+	if *systemdFlag {
+		listeners, err := activation.Listeners()
+		if err != nil {
+			log.Fatalf("error getting systemd listeners: %v", err)
+		}
+		for _, l := range listeners {
+			go connToChan(l, connChan)
+		}
+	} else {
+		user, err := os.UserCacheDir()
+		if err != nil {
+			log.Fatalf("error getting user cache directory: %v", err)
+		}
 
-	dir := filepath.Join(user, "sigstore", "gitsign")
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		log.Fatalf("error creating %s: %v", dir, err)
-	}
+		dir := filepath.Join(user, "sigstore", "gitsign")
+		if err := os.MkdirAll(dir, 0700); err != nil {
+			log.Fatalf("error creating %s: %v", dir, err)
+		}
 
-	path := filepath.Join(dir, "cache.sock")
-	if _, err := os.Stat(path); err == nil {
-		os.Remove(path)
-	}
-	fmt.Println(path)
+		path := filepath.Join(dir, "cache.sock")
+		if _, err := os.Stat(path); err == nil {
+			os.Remove(path)
+		}
+		fmt.Println(path)
 
-	l, err := net.Listen("unix", path)
-	if err != nil {
-		log.Fatalf("error opening socket: %v", err)
+		l, err := net.Listen("unix", path)
+		if err != nil {
+			log.Fatalf("error opening socket: %v", err)
+		}
+		go connToChan(l, connChan)
 	}
 	srv := rpc.NewServer()
 	if err := srv.Register(service.NewService()); err != nil {
 		log.Fatalf("error registering RPC service: %v", err)
 	}
+	for conn := range connChan {
+		go srv.ServeConn(conn)
+	}
+}
+
+func connToChan(l net.Listener, connChan chan net.Conn) {
 	for {
-		srv.Accept(l)
+		conn, err := l.Accept()
+		if err != nil {
+			log.Fatalf("error accepting connection: %v", err)
+		}
+		connChan <- conn
 	}
 }
